+++ Jolt/Shaders/ShaderWrapperCreator.h
@@ -90,7 +90,7 @@ public:
 		// When a constant buffer is bound, copy the data into the members
 		#undef JPH_SHADER_CONSTANTS_BEGIN
 		#undef JPH_SHADER_CONSTANTS_MEMBER
-		#define JPH_SHADER_CONSTANTS_BEGIN(type, name)	case HashString(#name): memcpy(&name + 1, inData, inSize);	break;
+		#define JPH_SHADER_CONSTANTS_BEGIN(type, name)	case HashString(#name): memcpy(&name + 1, inData, size_t(inSize));	break;
 		#define JPH_SHADER_CONSTANTS_MEMBER(type, name)
 
 		// When a buffer is bound, set the pointer
